+++ precompiles/erc20/utils_test.go
@@ -1,14 +1,18 @@
 package erc20_test
 
 import (
+	"math/big"
 	"time"
 
 	cryptotypes "github.com/cosmos/cosmos-sdk/crypto/types"
-
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/x/authz"
 	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
+	"github.com/ethereum/go-ethereum/accounts/abi"
+	"github.com/evmos/evmos/v15/precompiles/erc20"
 	commonfactory "github.com/evmos/evmos/v15/testutil/integration/common/factory"
+	utiltx "github.com/evmos/evmos/v15/testutil/tx"
+	erc20types "github.com/evmos/evmos/v15/x/erc20/types"
 )
 
 // setupSendAuthz is a helper function to set up a SendAuthorization for
@@ -39,6 +43,45 @@ func (s *PrecompileTestSuite) setupSendAuthz(
 	s.Require().NoError(err, "failed to execute MsgGrant")
 }
 
+// requireOut is a helper utility to reduce the amount of boilerplate code in the query tests.
+//
+// It requires the output bytes and error to match the expected values. Additionally, the method outputs
+// are unpacked and the first value is compared to the expected value.
+//
+// NOTE: It's sufficient to only check the first value because all methods in the ERC20 precompile only
+// return a single value.
+func (s *PrecompileTestSuite) requireOut(
+	bz []byte,
+	err error,
+	method abi.Method,
+	expPass bool,
+	errContains string,
+	expValue interface{},
+) {
+	if expPass {
+		s.Require().NoError(err, "expected no error")
+		s.Require().NotEmpty(bz, "expected bytes not to be empty")
+
+		// Unpack the name into a string
+		out, err := method.Outputs.Unpack(bz)
+		s.Require().NoError(err, "expected no error unpacking")
+
+		// Check if expValue is a big.Int. Because of a difference in uninitialized/empty values for big.Ints,
+		// this comparison is often not working as expected, so we convert to Int64 here and compare those values.
+		bigExp, ok := expValue.(*big.Int)
+		if ok {
+			bigOut, ok := out[0].(*big.Int)
+			s.Require().True(ok, "expected output to be a big.Int")
+			s.Require().Equal(bigExp.Int64(), bigOut.Int64(), "expected different value")
+		} else {
+			s.Require().Equal(expValue, out[0], "expected different value")
+		}
+	} else {
+		s.Require().Error(err, "expected error")
+		s.Require().Contains(err.Error(), errContains, "expected different error")
+	}
+}
+
 // requireSendAuthz is a helper function to check that a SendAuthorization
 // exists for a given grantee and granter combination for a given amount.
 //
@@ -64,4 +107,20 @@ func (s *PrecompileTestSuite) requireSendAuthz(grantee, granter sdk.AccAddress, 
 	} else {
 		s.Require().Equal(allowList, sendAuthz.AllowList, "expected different allow list")
 	}
+}
+
+// setupERC20Precompile is a helper function to set up an instance of the ERC20 precompile for
+// a given token denomination.
+func (s *PrecompileTestSuite) setupERC20Precompile(denom string) *erc20.Precompile {
+	tokenPair := erc20types.NewTokenPair(utiltx.GenerateAddress(), denom, erc20types.OWNER_MODULE)
+
+	precompile, err := erc20.NewPrecompile(
+		tokenPair,
+		s.network.App.BankKeeper,
+		s.network.App.AuthzKeeper,
+		s.network.App.TransferKeeper,
+	)
+	s.Require().NoError(err, "failed to create erc20 precompile")
+
+	return precompile
 }